Reorganiza las carpetas del proyecto
